refactor and clean up, fix counter in deluge_purge

+++ deluge_purge.py
@@ -19,12 +19,12 @@ ls_output = os.listdir(directory)
 os.chdir(directory)
 deleted = 0
 for each in ls_output:
-	if each not in active and each != 'temp' and each != '.htaccess' and each !='audio_sync':
+	if each not in active and each != 'temp' and each != '.htaccess' and each !='audio_sync' and each !='audio_completed':
 		print each
+		deleted += 1
 		if os.path.isfile(each):
 			os.remove(each)
 		elif os.path.isdir(each):
 			shutil.rmtree(each)
-		deleted += deleted
 
 print 'deleted',deleted, 'orphaned torrents'
